requirements.txt
networkx
numpy
pandas
scikit-learn
tensorflow
matplotlib
pyyaml